CPU percentages now have decimal precision

+++ userspace/libsanalyzer/analyzer.h
@@ -193,11 +193,11 @@ VISIBILITY_PRIVATE
 	//
 	uint64_t m_old_global_total_jiffies;
 	sinsp_procfs_parser* m_procfs_parser;
-	vector<uint32_t> m_cpu_loads;
-	vector<uint32_t> m_cpu_idles;
-	vector<uint32_t> m_cpu_steals;
+	vector<double> m_cpu_loads;
+	vector<double> m_cpu_idles;
+	vector<double> m_cpu_steals;
 	// Sum of the cpu usage of all the processes
-	uint32_t m_total_process_cpu;
+	double m_total_process_cpu;
 
 	//
 	// The table of aggreagted connections
